requirements.txt
matplotlib==3.8.0
numpy==2.2.4
pandas==1.5.3
pillow==9.4.0
scikit-learn==1.6.1
seaborn==0.13.2
torch==2.1.0
torchvision==0.16.0
